requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.1
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
more-itertools==9.0.0
multidict==6.0.2
psycopg2==2.9.4
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.42
typing_extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
zipp==3.9.0

